Fix a bug where only one piece of T-51 armor could spawn on raiders
It turns out the engine will only use one copy of each script in a quest, so only the last one was actually applying. To fix it, the script for injecting armor pieces has been rewritten to take an array of injection data to allow it to be used for all of the injections in a single set of power armor.