Attach worn legendary power armor mods using a magic effect instead of an actor script
Actor scripts can't be applied through templates that are referenced through leveled NPC lists, which means that all of the base NPCs that refer to those lists need to directly refer to the script template. This is bad for compatibility, especially for additional plugins that are only adding new enemy types to existing leveled NPC lists. Spells, on the other hand, can be templated through the leveled NPC lists.