Fix a bug where injections may not be registered on first install
The injection scripts were previously using OnInit() to inject, which does not require the OnInit() of any script they call to complete. As a result, there was a race condition where the array in the Injection Manager may not have been initialized yet. OnQuestInit() fires just after OnInit(), providing the needed functionality for injection, but it waits for OnInit() to be called in any script it calls into so we can guarantee that the array in the Injection Manager gets created before we register our injection.