Fix a Gunner on the Bradburton Overpass
The Gunner had a partial set of power armor linked to him, so he could spawn in power armor and then get out of it to get into the partial, rusted set. The Gunner can no longer spawn in his own power armor set, and the abandoned set is now a Gunner set.