Fix an issue where the PARTS patcher couldn't run with proxies on large load orders
The form ID calculation wasn't working - GetRecordById indicates it should be a form ID local to the file, but was always retrieving forms from Fallout.esm. The math to apply the load order doesn't work for large numbers (e.g. 237), but the editor ID works on large and small load orders.