Fix an issue where Maxson had two sets of identical power armor during Airship Down
Because Maxson was already in power armor, when he was forced into the power armor for the final quest, it would double up the items. To compensate, when the final battle quest is starting, Maxson is removed from his normal power armor before being moved into the final battle version.