Update some headers to 1.0
This probably does nothing. Issues were reported with these plugins, but they all turned out to be related to the user's setup, not this mod. However, other parts of the mod already use a header version of 1.0, so the older version of the game that only supports 0.95 isn't valid anyway - there's no harm in updating.